+++ example/grid/cell_stretch_points.py
@@ -8,10 +8,17 @@ points, stream = grid_points(
     y_spacing = 5
 )
 
+example_points = (
+    cq.Workplane("XY")
+    .pushPoints(stream)
+    .box(1,1,1)
+)
+
+
 cell_points = cell_stretch_points(
     points,
-    x_stretch = 2,
-    y_stretch = 2
+    x_stretch = 1,
+    y_stretch = 1
 )
 
 pattern = cq.Workplane("XY")
@@ -20,4 +27,7 @@ for points in cell_points:
     face = cq.Workplane("XY").polyline(points).close()
     pattern = pattern.add(face)
 
-#show_object(pattern)
+show_object(pattern)
+show_object(example_points)
+
+log(cell_points)