+++ src/cadqueryhelper/grid/grid_cell_basic.py
@@ -1,14 +1,16 @@
 import cadquery as cq
+from .cell_single import cell_single
 
 def grid_cell_basic(
-    cells_collection,
-    height:float=1,
-    taper=None
+    cells_collection:list[list[tuple[float,float]]],
+    height:float|None=1,
+    taper:float|None=None,
+    offset:float|None=None
 )->cq.Workplane:
     pattern = cq.Workplane("XY")
     
     for points in cells_collection:
-        face = cq.Workplane("XY").polyline(points).close().extrude(height,taper=taper)
+        face = cell_single(points, height, taper, offset)
         pattern = pattern.add(face)
         
     return pattern